requirements.txt
fastapi
uvicorn
rembg==2.0.68
onnxruntime==1.16.3
pillow
numpy
scikit-image
